# docs/openapi.yaml
openapi: 3.0.0
paths:
  /api/v1/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login avvenuto con successo, oppure richiesta MFA
        '401':
          description: Credenziali errate o account non attivato
        '429':
          description: Troppe richieste, riprovare più tardi
      summary: Login utente e generazione token JWT
      tags:
        - Auth
  /api/v1/auth/mfa-verify:
    post:
      operationId: AuthController_mfaVerify
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyDto'
      responses:
        '200':
          description: MFA verificato con successo
        '401':
          description: Token temporaneo scaduto o codice MFA non valido
        '429':
          description: Troppe richieste, riprovare più tardi
      summary: Verifica codice MFA per completare il login
      tags:
        - Auth
  /api/v1/auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      responses:
        '200':
          description: Nuovo access token generato
        '401':
          description: Refresh token assente, non valido o scaduto
      summary: Rinnova l'access token usando il refresh token opaco (cookie rtk)
      tags:
        - Auth
...
  /api/v1/auth/activate:
    post:
      operationId: AuthController_activate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateAccountDto'
      responses:
        '200':
          description: Account attivato e password impostata
        '400':
          description: Password non conforme alla policy
        '401':
          description: Token non valido o scaduto
        '429':
          description: Troppe richieste, riprovare più tardi
      summary: Attiva un account utente tramite token e imposta la password
      tags:
        - Auth
  /api/v1/auth/forgot-password:
    post:
      operationId: AuthController_forgotPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        '200':
          description: Richiesta elaborata (risposta generica per sicurezza)
        '429':
          description: Troppe richieste, riprovare più tardi
      summary: Richiede il recupero password inviando una email con token
      tags:
        - Auth
  /api/v1/auth/reset-password:
    post:
      operationId: AuthController_resetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Password reimpostata con successo
        '400':
          description: Password non conforme alla policy
        '401':
          description: Token non valido o scaduto
        '429':
          description: Troppe richieste, riprovare più tardi
      summary: Reimposta la password tramite token di recupero
      tags:
        - Auth
  /api/v1/auth/request-activation:
    post:
      operationId: AuthController_requestActivation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        '200':
          description: Email di attivazione inviata con successo
        '400':
          description: Account già attivo
        '404':
          description: Utente non trovato
      summary: Richiede o re-invia l'email di attivazione per un utente (Admin+)
      tags:
        - Auth
  /api/v1/auth/me:
    get:
      operationId: AuthController_getMe
      parameters: []
      responses:
        '200':
          description: Dati utente recuperati
      summary: Recupera i dati dell'utente autenticato
      tags:
        - Auth
    patch:
      operationId: AuthController_updateProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Dati anagrafici aggiornati con successo
        '400':
          description: Nome o cognome non validi
      summary: Aggiorna nome e cognome dell'utente autenticato (pagina profilo)
      tags:
        - Auth
  /api/v1/auth/change-password:
    patch:
      operationId: AuthController_changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
      responses:
        '200':
          description: Password cambiata con successo
        '400':
          description: Nuova password non conforme alla policy
        '401':
          description: Password attuale non corretta
      summary: Cambia la password dell'utente autenticato (pagina profilo)
      tags:
        - Auth
...
  /api/v1/auth/mfa-enable:
    post:
      operationId: AuthController_mfaEnable
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaEnableDto'
      responses:
        '200':
          description: MFA abilitata con successo
        '400':
          description: MFA già abilitata, setup scaduto o codice non valido
      summary: Abilita la MFA per l'utente autenticato
      tags:
        - Auth
  /api/v1/auth/mfa-disable:
    post:
      operationId: AuthController_mfaDisable
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaDisableDto'
      responses:
        '200':
          description: MFA disabilitata con successo
        '400':
          description: MFA non abilitata o codice non valido
      summary: Disabilita la MFA per l'utente autenticato
      tags:
        - Auth
...
  /api/v1/auth/sessions/{sessionId}:
    delete:
      operationId: AuthController_revokeSession
      parameters:
        - name: sessionId
          required: true
          in: path
          description: Identificativo opaco della sessione (16 char hex)
          schema:
            type: string
      responses:
        '200':
          description: Sessione revocata con successo
        '404':
          description: Sessione non trovata o non appartenente all'utente
      summary: Revoca una sessione/dispositivo dell'utente autenticato
      tags:
        - Auth
  /api/v1/auth/impersonate/{guid}:
    post:
      operationId: AuthController_impersonate
      parameters:
        - name: guid
          required: true
          in: path
          description: Guid pubblico (16 char) dell'utente da impersonare
          schema:
            type: string
      responses:
        '200':
          description: Access token di impersonificazione generato
        '403':
          description: Permessi insufficienti o target è un SuperAdmin/utente disabilitato
        '404':
          description: Utente target non trovato
      summary: Avvia l'impersonificazione di un utente (SuperAdmin only)
      tags:
        - Auth
  /api/v1/auth/end-impersonation:
    post:
      operationId: AuthController_endImpersonation
      parameters: []
      responses:
        '200':
          description: Sessione del SuperAdmin originale ripristinata
        '400':
          description: Nessuna impersonificazione in corso
        '401':
          description: Sessione del SuperAdmin originale non più valida
      summary: >-
        Termina l'impersonificazione corrente e ripristina la sessione del
        SuperAdmin originale
      tags:
        - Auth
...
info:
  title: Starter Kit API
  description: >-
    API REST del boilerplate aziendale (auth, RBAC, MFA, audit log, gestione
    utenti).
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    LoginDto:
      type: object
      properties:
        email:
          type: string
          description: Indirizzo email dell'utente
          example: admin@example.com
        password:
          type: string
          description: Password dell'utente
          example: Password123!
      required:
        - email
        - password
    MfaVerifyDto:
      type: object
      properties:
        tmpToken:
          type: string
          description: Token temporaneo restituito da /auth/login quando è richiesta la MFA
          example: a1b2c3...
        code:
          type: string
          description: Codice TOTP a 6 cifre generato dall'app authenticator
          example: '123456'
      required:
        - tmpToken
        - code
    ActivateAccountDto:
      type: object
      properties:
        token:
          type: string
          description: Il token di attivazione ricevuto via email (64 caratteri)
          example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6a1b2c3d4e5f6
        password:
          type: string
          description: La nuova password da impostare per l'account (minimo 12 caratteri)
          example: SicuraPassword123!
      required:
        - token
        - password
    ForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
          description: L'indirizzo email associato all'account da recuperare
          example: utente@example.com
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        token:
          type: string
          description: Il token di reimpostazione ricevuto via email (64 caratteri)
          example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6a1b2c3d4e5f6
        password:
          type: string
          description: >-
            La nuova password conforme alla policy di sicurezza (minimo 12
            caratteri)
          example: NuovaSuperPassword2026!
      required:
        - token
        - password
    UpdateProfileDto:
      type: object
      properties:
        name:
          type: string
          description: Nome dell'utente
          example: Mario
        surname:
          type: string
          description: Cognome dell'utente
          example: Rossi
      required:
        - name
    ChangePasswordDto:
      type: object
      properties:
        currentPassword:
          type: string
          description: La password attuale dell'utente, usata per confermare l'identità
          example: VecchiaPassword2025!
        newPassword:
          type: string
          description: >-
            La nuova password conforme alla policy di sicurezza (minimo 12
            caratteri)
          example: NuovaSuperPassword2026!
      required:
        - currentPassword
        - newPassword
    MfaEnableDto:
      type: object
      properties:
        code:
          type: string
          description: >-
            Codice TOTP a 6 cifre generato dall'app authenticator con il secret
            ricevuto da mfa-setup
          example: '123456'
      required:
        - code
    MfaDisableDto:
      type: object
      properties:
        code:
          type: string
          description: Codice TOTP a 6 cifre generato dall'app authenticator
          example: '123456'
      required:
        - code
...